+++ views/professional.jade
@@ -6,12 +6,12 @@ block content
     .col-md-12
       table.table
         th ID
+        th League
         th Radiant
         th Dire
-        th League
         each match in matches
           tr
-            td=match.match_id
+            td: a(href="/matches/"+match.match_id)=match.match_id
             td=match.leagueid
             td=match.radiant_name
             td=match.dire_name
